Data.MonoidMap

Description

Total monoidal map type with support for semigroup and monoid subclasses.

Type

data MonoidMap k v

Construction

empty :: MonoidMap k v

The empty MonoidMap.

fromList :: (Ord k, MonoidNull v) => [(k, v)] -> MonoidMap k v

Constructs a MonoidMap from a list of key-value pairs.

If the list contains more than one value for the same key, values are combined together with <>.

fromListWith

Arguments

:: (Ord k, MonoidNull v)

=> (v -> v -> v)
Combination function with which to combine values for duplicate keys.

-> [(k, v)]

-> MonoidMap k v

Constructs a MonoidMap from a list of key-value pairs.

If the list contains more than one value for the same key, values are combined together with the given combination function.

fromMap :: MonoidNull v => Map k v -> MonoidMap k v

Constructs a MonoidMap from an ordinary Map.

singleton :: (Ord k, MonoidNull v) => k -> v -> MonoidMap k v

Constructs a MonoidMap from a single key-value pair.

Deconstruction

toList :: MonoidMap k v -> [(k, v)]

Converts a MonoidMap to a list of key-value pairs.

The result only includes entries with values that are not null.

toMap :: MonoidMap k v -> Map k v

Converts a MonoidMap to a Map.

The result only includes entries with values that are not null.

Basic operations

get :: (Ord k, Monoid v) => k -> MonoidMap k v -> v

Gets the value associated with the given key.

set :: (Ord k, MonoidNull v) => k -> v -> MonoidMap k v -> MonoidMap k v

Sets the value associated with the given key.

adjust :: (Ord k, MonoidNull v) => (v -> v) -> k -> MonoidMap k v -> MonoidMap k v

Adjusts the value associated with the given key.

delete :: (Ord k, MonoidNull v) => k -> MonoidMap k v -> MonoidMap k v

Sets the value associated with the given key to mempty.

Queries

keys :: MonoidMap k v -> Set k

Returns the set of keys associated with values that are not null.

member :: Ord k => k -> MonoidMap k v -> Bool

Returns True if (and only if) the given key is associated with a value that is not null.

notMember :: Ord k => k -> MonoidMap k v -> Bool

Returns True if (and only if) the given key is associated with a value that is null.

null :: MonoidMap k v -> Bool

Returns True if (and only if) all values in the map are null.

notNull :: MonoidMap k v -> Bool

Returns True if (and only if) the map contains at least one value that is not null.

size :: MonoidMap k v -> Int

Returns a count of all values in the map that are not null.

Slicing

take :: Int -> MonoidMap k v -> MonoidMap k v

Takes a given number of entries in key order, beginning with the smallest keys.

take n == fromList . take n . toList

drop :: Int -> MonoidMap k v -> MonoidMap k v

Drops a given number of entries in key order, beginning with the smallest keys.

drop n == fromList . drop n . toList

splitAt :: Int -> MonoidMap k a -> (MonoidMap k a, MonoidMap k a)

Splits a map at a particular index.

splitAt n xs == (take n xs, drop n xs)

Filtering

filter :: (k -> v -> Bool) -> MonoidMap k v -> MonoidMap k v

Filters a map according to a predicate on keys and values.

The result contains just the subset of entries that satisfy the predicate.

filterKeys :: (k -> Bool) -> MonoidMap k v -> MonoidMap k v

Filters a map according to a predicate on keys.

The result contains just the subset of entries that satisfy the predicate.

filterKeys f m == filter (k _ -> f k) m

filterValues :: (v -> Bool) -> MonoidMap k v -> MonoidMap k v

Filters a map according to a predicate on values.

The result contains just the subset of entries that satisfy the predicate.

filterValues f m == filter (_ v -> f v) m

Partitioning

partition :: (k -> v -> Bool) -> MonoidMap k v -> (MonoidMap k v, MonoidMap k v)

Partitions a map according to a predicate on keys and values.

The first map contains all entries that satisfy the predicate, and the second map contains all entries that fail the predicate.

partition f m == (filter f m, filter (k v -> not (f k v)) m)

partitionKeys :: (k -> Bool) -> MonoidMap k v -> (MonoidMap k v, MonoidMap k v)

Partitions a map according to a predicate on keys.

The first map contains all entries that satisfy the predicate, and the second map contains all entries that fail the predicate.

partitionKeys f m == (filterKeys f m, filterKeys (not . f) m)

partitionValues :: (v -> Bool) -> MonoidMap k v -> (MonoidMap k v, MonoidMap k v)

Partitions a map according to a predicate on values.

The first map contains all entries that satisfy the predicate, and the second map contains all entries that fail the predicate.

partitionValues f m == (filterValues f m, filterValues (not . f) m)

Traversal

map :: MonoidNull v2 => (v1 -> v2) -> MonoidMap k v1 -> MonoidMap k v2

Applies the given function to all values in the map that are not null.

If the range of the given function includes mempty, then the resultant map may have a size that is smaller than the original map.